django/requirements.txt
pip==20.0.2
Django==1.11.6
psycopg2==2.8.4
djangorestframework==3.9.3
django-cors-headers==2.4.0
